refactor(s2n-quic-dc): put map impl behind trait
